% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockTable.R
\name{stockTableUI}
\alias{stockTableUI}
\title{按纽生成生成器界面}
\usage{
stockTableUI(
  tabTitle = "查看即时库存更新报表",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = stockTableUI_left,
  func_right = stockTableUI_right,
  func_bottom = stockTableUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
按纽生成生成器界面
}
\examples{
stockTableUI()
}
